Add empty page component
Added component for when there are no suggestions to show. Only styled
for mobile. Removed button component and replaced it with button and
span jsx. Ran into issues with scss modules conflicting with each other
when used in different modules. It's only three lines of code, so I just
included it within the components instead of extracting to a separate
component.