Add background modal for SuggestionPage
When the sort dropdown is open on the Suggestion page, if the user
wanted to close the dropdown by clicking on the body it would fire the
click event listeners of the suggestion boxes and send the user to
another page. With the modal as a transparent background, the modal acts
as a way of closing the dropdown when it's clicked on.