Add keyboard user detection function
ISSUE: elemnents with a tabindex assigned to them have an oultine when
focused by touch or click

SOLVED: created and initiated a function to add or remove styles based
on whether or not the focus is from the keyboard or mouse click. Now the
outline styles only appear on tabindexed elements when focused on by the
keyboard.