Fix issue of unwanted rerender w/ wrong sort
ISSUE: when user clicks the like button was clicked, the request list
would automatically sort in an incorrect order, which was the default
sort order

FIX: Sent sortype along with the dispatch action object and the requests
are sorted based on its value. Sorting occurs in the reducer now, too.