Refactor request reducer/SuggestionList, bug fix
Fixed an issue with the ordering of requests when compared but have the
same value. Now there's a conditional to check if the value are the same
in order to keep sorted ordr in place. Also refactored the request
reducer and suggestion list component to compare numbers rather than
strings and implemented ternary conditionals