Add user ability to reply to replies
Changed the structure of data.json to a flat structure rather than a
nested sructure when it comes to the comments and replies. They both now
share the same comments array and assigned top level comments a parentID
of null and replies have the parentID of the comments. Used a recursive
component render with the comments component in order to render out
comments and replies.