Add currentUser check to display edit btn
Added a new author property to the ProductRequest class and data
structure of data.json. This property is used to check if the current
suggestion/request is from the currentUser that's signed in or not. If
it is, an edit button gets displayed for every suggestion/request that
the currentUser creates alowwing them to edit thier feedback, otherwise
the edit button does not render